migrate ConsumerReportClosedNoReading
