+++ app/services/emails/EmailDefinitionsConsumer.scala
@@ -349,6 +349,44 @@ object EmailDefinitionsConsumer {
 
   }
 
+  case object ConsumerReportClosedNoReading extends EmailDefinition {
+    override val category = Consumer
+
+    override def examples =
+      Seq(
+        "report_closed_no_reading" -> ((recipient, messagesApi) =>
+          EmailImpl(genReport.copy(email = recipient), Some(genCompany), messagesApi)
+        ),
+        "report_closed_no_reading_case_dispute" -> ((recipient, messagesApi) =>
+          EmailImpl(
+            genReport.copy(email = recipient, tags = List(ReportTag.LitigeContractuel)),
+            Some(genCompany),
+            messagesApi
+          )
+        )
+      )
+
+    final case class EmailImpl(
+        report: Report,
+        maybeCompany: Option[Company],
+        messagesApi: MessagesApi
+    ) extends Email {
+      private val lang                                        = Lang(getLocaleOrDefault(report.lang))
+      implicit private val messagesProvider: MessagesProvider = MessagesImpl(lang, messagesApi)
+
+      override val recipients: List[EmailAddress] = List(report.email)
+      override val subject: String                = messagesApi("ReportClosedByNoReadingEmail.subject")(lang)
+
+      override def getBody: (FrontRoute, EmailAddress) => String = (frontRoute, _) =>
+        views.html.mails.consumer.reportClosedByNoReading(report, maybeCompany)(frontRoute, messagesProvider).toString
+
+      override def getAttachements: AttachmentService => Seq[Attachment] =
+        _.needWorkflowSeqForWorkflowStepN(3, report)
+
+    }
+
+  }
+
   private def getLocaleOrDefault(locale: Option[Locale]): Locale = locale.getOrElse(Locale.FRENCH)
 
 }
